refactor: split plan consume service
