+++ arex-schedule-web-api/src/main/java/com/arextest/schedule/service/PlanConsumePrepareService.java
@@ -1,0 +1,158 @@
+package com.arextest.schedule.service;
+
+import com.arextest.schedule.bizlog.BizLogger;
+import com.arextest.schedule.common.CommonConstant;
+import com.arextest.schedule.dao.mongodb.ReplayActionCaseItemRepository;
+import com.arextest.schedule.dao.mongodb.ReplayPlanRepository;
+import com.arextest.schedule.model.*;
+import com.arextest.schedule.planexecution.PlanExecutionContextProvider;
+import com.arextest.schedule.progress.ProgressEvent;
+import com.arextest.schedule.utils.ReplayParentBinder;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.collections4.CollectionUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.stereotype.Service;
+
+import javax.annotation.Resource;
+import java.util.List;
+import java.util.Set;
+
+/**
+ * Created by Qzmo on 2023/7/5
+ */
+@Service
+@Slf4j
+@SuppressWarnings({"rawtypes", "unchecked"})
+public class PlanConsumePrepareService {
+    @Resource
+    private MetricService metricService;
+    @Resource
+    private ReplayPlanRepository replayPlanRepository;
+    @Resource
+    private ReplayReportService replayReportService;
+    @Resource
+    private ReplayCaseRemoteLoadService caseRemoteLoadService;
+    @Resource
+    private PlanExecutionContextProvider planExecutionContextProvider;
+    @Resource
+    private ProgressEvent progressEvent;
+    @Resource
+    private ReplayActionCaseItemRepository replayActionCaseItemRepository;
+
+    public int prepareRunData(ReplayPlan replayPlan) {
+        metricService.recordTimeEvent(LogType.PLAN_EXECUTION_DELAY.getValue(), replayPlan.getId(), replayPlan.getAppId(), null,
+                System.currentTimeMillis() - replayPlan.getPlanCreateMillis());
+        int planSavedCaseSize = saveAllActionCase(replayPlan.getReplayActionItemList());
+        if (planSavedCaseSize != replayPlan.getCaseTotalCount()) {
+            LOGGER.info("update the plan TotalCount, plan id:{} ,appId: {} , size: {} -> {}", replayPlan.getId(),
+                    replayPlan.getAppId(), replayPlan.getCaseTotalCount(), planSavedCaseSize);
+            replayPlan.setCaseTotalCount(planSavedCaseSize);
+            replayPlanRepository.updateCaseTotal(replayPlan.getId(), planSavedCaseSize);
+            replayReportService.updateReportCaseCount(replayPlan);
+        }
+        return planSavedCaseSize;
+    }
+
+    private int saveAllActionCase(List<ReplayActionItem> replayActionItemList) {
+        int planSavedCaseSize = 0;
+        long start;
+        long end;
+
+        for (ReplayActionItem replayActionItem : replayActionItemList) {
+            if (replayActionItem.getReplayStatus() != ReplayStatusType.INIT.getValue()) {
+                planSavedCaseSize += replayActionItem.getReplayCaseCount();
+                continue;
+            }
+            int preloaded = replayActionItem.getReplayCaseCount();
+
+            start = System.currentTimeMillis();
+            int actionSavedCount = streamingCaseItemSave(replayActionItem);
+            end = System.currentTimeMillis();
+            BizLogger.recordActionItemCaseSaved(replayActionItem, actionSavedCount, end - start);
+
+            planSavedCaseSize += actionSavedCount;
+            if (preloaded != actionSavedCount) {
+                replayActionItem.setReplayCaseCount(actionSavedCount);
+                LOGGER.warn("The saved case size of actionItem not equals, preloaded size:{},saved size:{}", preloaded,
+                        actionSavedCount);
+            }
+            progressEvent.onActionCaseLoaded(replayActionItem);
+        }
+        return planSavedCaseSize;
+    }
+
+    private int streamingCaseItemSave(ReplayActionItem replayActionItem) {
+        List<ReplayActionCaseItem> caseItemList = replayActionItem.getCaseItemList();
+        int size;
+        if (CollectionUtils.isNotEmpty(caseItemList)) {
+            size = doFixedCaseSave(caseItemList);
+        } else {
+            size = doPagingLoadCaseSave(replayActionItem);
+        }
+        return size;
+    }
+
+
+    private void caseItemPostProcess(List<ReplayActionCaseItem> caseItemList) {
+        // to provide necessary fields into case item for context to consume when sending
+        planExecutionContextProvider.injectContextIntoCase(caseItemList);
+    }
+
+    /**
+     * Paging query storage's recording data.
+     * if caseCountLimit > CommonConstant.MAX_PAGE_SIZE, Calculate the latest pageSize and recycle pagination queries
+     * <p>
+     * else if caseCountLimit < CommonConstant.MAX_PAGE_SIZE or recording data size < request page size,
+     * Only need to query once by page
+     */
+    private int doPagingLoadCaseSave(ReplayActionItem replayActionItem) {
+        final ReplayPlan replayPlan = replayActionItem.getParent();
+        long beginTimeMills = replayActionItem.getLastRecordTime();
+        if (beginTimeMills == 0) {
+            beginTimeMills = replayPlan.getCaseSourceFrom().getTime();
+        }
+        long endTimeMills = replayPlan.getCaseSourceTo().getTime();
+        int totalSize = 0;
+        int caseCountLimit = replayActionItem.getOperationTypes() == null ? replayPlan.getCaseCountLimit() : replayPlan.getCaseCountLimit() * replayActionItem.getOperationTypes().size();
+        int pageSize = Math.min(caseCountLimit, CommonConstant.MAX_PAGE_SIZE);
+        while (beginTimeMills < endTimeMills) {
+            List<ReplayActionCaseItem> caseItemList = caseRemoteLoadService.pagingLoad(beginTimeMills, endTimeMills,
+                    replayActionItem, caseCountLimit - totalSize);
+            if (CollectionUtils.isEmpty(caseItemList)) {
+                break;
+            }
+            caseItemPostProcess(caseItemList);
+            ReplayParentBinder.setupCaseItemParent(caseItemList, replayActionItem);
+            totalSize += caseItemList.size();
+            beginTimeMills = caseItemList.get(caseItemList.size() - 1).getRecordTime();
+            replayActionCaseItemRepository.save(caseItemList);
+            if (totalSize >= caseCountLimit || caseItemList.size() < pageSize) {
+                break;
+            }
+            replayActionItem.setLastRecordTime(beginTimeMills);
+        }
+        return totalSize;
+    }
+
+    private int doFixedCaseSave(List<ReplayActionCaseItem> caseItemList) {
+        int size = 0;
+        for (int i = 0; i < caseItemList.size(); i++) {
+            ReplayActionCaseItem caseItem = caseItemList.get(i);
+            Set<String> operationTypes = caseItem.getParent().getOperationTypes();
+            ReplayActionCaseItem viewReplay = caseRemoteLoadService.viewReplayLoad(caseItem, operationTypes);
+            if (viewReplay == null) {
+                caseItem.setSendStatus(CaseSendStatusType.REPLAY_CASE_NOT_FOUND.getValue());
+                caseItem.setSourceResultId(StringUtils.EMPTY);
+                caseItem.setTargetResultId(StringUtils.EMPTY);
+            } else {
+                viewReplay.setParent(caseItem.getParent());
+                caseItemList.set(i, viewReplay);
+                size++;
+            }
+        }
+
+        caseItemPostProcess(caseItemList);
+        replayActionCaseItemRepository.save(caseItemList);
+        return size;
+    }
+}
